change route to /home
