+++ frontend/src/components/Home/index.js
@@ -1,0 +1,72 @@
+import useFetch from "@/hooks/useFetch";
+import QuickLinksAndLatestUpdates from "@/components/QuickLinksandLatestUpdates";
+import Announcements from "@/components/Annocements";
+import FAQ from "@/components/FAQ";
+import ImageGallery from "@/components/Home/ImageGallery/ImageGallery";
+import Achievements from "@/components/Home/Achievements/Card";
+import Carousel from "@/components/Home/Events/Carousel/Carousel";
+import { Manrope } from "next/font/google";
+const manrope = Manrope({ subsets: ["latin"], weight: ["600", "400"] });
+
+export default function Home() {
+  const { data, error, loading } = useFetch(
+    "https://swc.iitg.ac.in/saPortal/api/homes?populate=deep"
+  );
+
+  return (
+    <>
+      {data && (
+        <>
+          <div className={manrope.className}>
+            <div className="w-full flex flex-col justify-center items-center gap-4 p-6 ">
+              <div className="w-[83vw]">
+                <div className="flex  gap-[20vw]">
+                  <div className="flex flex-col">
+                    <div className="text-3xl font-semibold">
+                      Welcome To The <br />
+                      Students Affairs Board
+                    </div>
+                    <div className="text-sm pt-6">
+                      {data.data[0].attributes.About}
+                    </div>
+                  </div>
+                  <div className=" hidden md:flex flex-col p-8 bg-[#FBFBFB] rounded-lg h-fit w-60">
+                    <div>
+                      <div className="text-gray font-semibold">Content</div>
+                    </div>
+                    <div className="border-l-2 px-2 text-xs border-gray-400 ">
+                      <div>Quick links</div>
+                      <div className="text-gray">Achievement</div>
+                      <div className="text-gray">Announcement</div>
+                    </div>
+                  </div>
+                </div>
+              </div>
+
+              <div className="w-[83vw]">
+                <QuickLinksAndLatestUpdates
+                  prop={data.data[0].attributes}
+                ></QuickLinksAndLatestUpdates>
+                <div className="flex flex-col w-full md:w-2/3">
+                  <div className="text-3xl font-semibold mb-8">
+                    Achievements
+                  </div>
+                  <div className=" columns-1 md:columns-2">
+                    {data.data[0].attributes.Achievements.map((item) => (
+                      <Achievements props={item}></Achievements>
+                    ))}
+                  </div>
+                  <Announcements prop={data.data[0].attributes}></Announcements>
+                  <FAQ prop={data.data[0].attributes}></FAQ>
+                </div>
+              </div>
+              {/* <Achievements prop={data.data[0].attributes}></Achievements> */}
+            </div>
+          </div>
+          <ImageGallery prop={data.data[0].attributes}></ImageGallery>
+          <Carousel prop={data.data[0].attributes}></Carousel>
+        </>
+      )}
+    </>
+  );
+}
