+++ frontend/src/components/QuickLinksandLatestUpdates.js
@@ -3,7 +3,7 @@ import QuickLinks from "./QuickLinks";
 const QuickLinksAndLatestUpdates = ({ prop }) => {
   return (
     <>
-      <div className="flex flex-col lg:flex-row">
+      <div className="flex flex-col lg:justify-between lg:flex-row">
         <QuickLinks prop={prop} />
         <LatestUpdates prop={prop} />
       </div>
